compiler-modules: Convert batch-driver to a module
For now, we use "batch-driver" as the module name, to make things
simple.  Originally, it looks like the idea was that you can have
other kinds of drivers, like an interactive one or a staged one,
or one used through the interpreter.  We can try to do this through
the use of functors, perhaps.   This makes things simpler for us:
Now we can just append .import.scm to the filename in the build
system.  The unit now also has the same name as the module, which
may be undone when we add a functor and an implementation selection
mechanism.

Add a few missing names to compiler-namespace which would otherwise be
undefined within batch-driver.

The apply-test was using "printf" in a macro, which happened to be
available when running inside the compiler due to namespace leakage,
which is now fixed through proper module usage.  It now cleanly
imports extras, which provides the printf procedure.