compiler-modules: Convert c-platform to a module.
A few variables (default-optimization-passes eq-inline-operator,
membership-test-operators and membership-unfold-limit) were defined by
c-platform but used by optimizer.  This would result in a circular
dependency, because c-platform is already using various things which
were exported by the optimizer.  For this reason, and to reduce
coupling between the optimizer and the platform-specific code,
optimizer now defines the variables, which are then set by the
specific platform in use.  This makes it more explicit what the
optimizer needs from a platform.

Removed the unused variable default-debugging-declarations.

Rewrite is now exported cleanly by optimizer, and used by c-platform,
so it could be removed from compiler-namespace.

Added small-parameter-limit to compiler-namespace, for the time being.