compiler-modules: Convert scrutinizer to a module.
In chicken-syntax, :, the, define-specialization, compiler-typecase,
and define-type now expand directly to
scrutinizer#[check-and-]validate-type, which is needed in order to
remove these two procedures from compiler-namespace.

Fixed lfa2 debugging code to look at lfa2-debug instead of
scrutiny-debug.  Converted use of string-concatenate to string-join to
avoid dependency on srfi-13, which is slated for extraction from core.