compiler-modules: Convert support to a module.
This is quite a big module which is used throughout the compiler, so
this is necessarily a very large change.  I basically only went through,
checked if and where identifiers were used and added them to the export
list if they were used elsewhere.  If they were only used in one place,
I added a small note to indicate it, for a future, proper cleanup.

In chicken-ffi-syntax.scm, the foreign definition macros now expand
directly to support#foreign-type->scrutiny-type, and in
chicken-syntax.scm, define-specialization now directly expands to
support#variable-mark.  This is needed in order to remove
foreign-type->scrutiny-type and variable-mark from compiler-namespace.

With this, a few things have been cleaned up:

- The "quit" procedure has been renamed to "quit-compiling", to avoid
confusion with "quit" from library.scm.
- The "words" procedure has been renamed "bytes->words", for improved clarity and consistency with words->bytes.
- The decompose-lambda-list alias has been removed in favor of using ##sys#decompose-lambda-list directly.  Possibly this could be reverted and the alias moved into library.scm.
- The "count!" and "pprint-expressions-to-file" procedures have been removed, as they were unused.
- The disabled-warnings list has been removed, as it was unused.
- slashify and uncommentify have been moved to c-backend, where they belong.