compiler-modules: Simplify compiler module import forms
Prefix the get, get-all, get-list and put! procedures with "db-".
This prevents confusion with the 2-argument "get" and "put!" versions
from library.scm, and it also cleans up the chicken module import.

Remove syntax-error from exclusion list: The module system merely
aliases identifiers, so it doesn't need any special handling in order
to prevent compiler modules from including the "wrong" syntax-error
procedure.  Add a test for the aliasing behaviour of the module
system.

The debugging-chicken variable is used in a similar fashion: it is
defined in and exported by support, and it is set! in batch-driver.
It is read by both support and compiler.