compiler-modules: Make compiler-arguments and source-file nonglobal by passing them around.
The compiler-arguments global was declared in chicken.scm and used
only in c-backend to write the compiler arguments to the resulting
C file, for informational/debugging/forensic purposes.  Just like
the filename and the final (derived) options, this is now passed
around from chicken to batch-driver, and from batch-driver to
c-backend.  The variable has been renamed for (hopefully) increased
clarity.

As stated above, the "filename" in batch-driver is passed to some
procedures.  However, it is not passed to all: emit-type-file and
emit-global-inline-file were the only two procedures to read the
filename from the global "source-filename".  This global has now
been removed and the file name is passed to the procedures as
an argument.