compiler-modules: Fix -no-parentheses-synonyms flag to compiler.
The patch in d31f8ea988425e94745258f174a86fdb2bb06459 was apparently
incomplete: it fixed the interpreter's handling of the
-no-parentheses-synonyms flag, but this flag was not used by the
compiler: it was listed amongst the valid compiler options with
arguments, which is wrong, and the batch-driver used it as a
no-argument flag, but with a typo in its name, so it would never
get used.  The version with typo is also used in compiler-namespace.