compiler-modules: Reduce coupling between components by passing around more options.
The options in question are used in a lot of places, but it is only
necessary to set them in batch-driver, and export them from compiler.
When invoking a subcomponent of the compiler, batch-driver can pass
the various options to that component.  This is inspired by the fact
that the scrutinizer's "scrutinize" procedure already received
"specialize" as an argument.  This corresponds to a global
"enable-specialization" option defined in compiler.  This idea has
been taken further, with the following options:
- enable-specialization (wasn't passed to load-type-database)
- block-compilation
- inline-locally
- inline-substitutions-enabled
- inline-max-size
- strict-variable-types