compiler-modules: Move foreign callback code from compiler to support.
Because support makes use of foreign-callback accessors, it makes more
sense to put the record definition there as well.  The foreign-type-table
also fits better in support; after moving it there we can put accessors
and setters, as well as an initialization procedure into a "proper" API
which the other modules can consume.

Foreign stubs are now registered through a separate API as well, so that
the foreign stub list can be moved to support.  Unfortunately, the
c-backend still accesses this list directly.