compiler-modules: Make bindings non-global.
This is a bit of a tricky change.  The reason is that the platform
defines the *default* standard, extended, internal and foldable bindings.
The final extended-bindings and standard-bindings, however, are possibly
overridden through declarations by the compiler.  The compiler takes the
bindings from the default bindings.

In order to make all this work without creating cyclic dependencies
between the backend/platform and the compiler, we've moved the definitions
of *all* bindings to the compiler.  The platform then set!s these to the
correct values.  This creates a new dependency between platform and
compiler, but that's okay (the backend also depends on the compiler, and
it exists at about the same level).  The compiler-syntax also makes use
of these bindings, so it makes sense to define a dependency for
compiler-syntax on the compiler.