Make import load, add unit info to import libraries, track unit execution
Adds a load flag to import expansion, to indicate whether to a module's
code should be loaded when its identifiers are imported.

Tracks library loading with a new "unit hook". This requires pulling
build-version into library.scm so that library (which provides the hook)
is always the first unit loaded.