Use core form to quote identifiers in expansion of `require-extension' et al.
This patch avoids including an unqualified `quote' in any of the forms
produced by `##sys#do-the-right-thing', which can cause failures in
modules that don't import the "scheme" module since a bare `quote' will
be unbound in the expansion of e.g. `require-extension'. Without it, the
resulting errors will generally look like this:

  Warning: reference to possibly unbound identifier `<library>`

  Warning: reference to possibly unbound identifier `quote'
  Warning:    suggesting one of:
  Warning:    (import r5rs-null)
  Warning:    (import r4rs-null)
  Warning:    (import scheme)
  Warning:    (import r4rs)`

Fixes #1403 and #1343.

Signed-off-by: Peter Bex <peter@more-magic.net>