Get rid of ##sys#globalize and a hack in ##core#declare
In ##core#declare, there was some questionable messing around with the
parameterization of ##sys#current-environment after which it called
process-declaration.  And process-declaration was the only procedure
which was still using ##sys#globalize so it can be dropped now.

The process of looking up variables in declarations can be a
simplified version of "resolve-variable" because it needs to take into
account only regular bindings.  So, we can just look it up in the
syntax env (resolving macro aliases) and then use the alias global
hook to module-prefix it, if needed.

Signed-off-by: felix <felix@call-with-current-continuation.org>