Replace rest ops with list ops on explicitly consed rest args
When the optimizer detects a "rest op" for a rest parameter which has
been replaced by an explicitly consed list at the call site, the rest
op is replaced by car/cdr/length list operations on the now-explicit
argument.  This is needed because the argvector will not contain those
extra arguments anymore, so attempting to read them is an access
outside the argvector's bounds.

This is sometimes detected by the runtime with an error like
"attempted rest argument access at index 0 but rest list length is 0",
but in other situations it will cause a segmentation fault, as has
been reported in #1756.

Signed-off-by: megane <meganeka@gmail.com>