Add new -:R runtime option to influence how srand() is called
Without this, there's some nondeterminism when running an
application, which defends against symbol table stuffing attacks
but can make debugging or benchmarking GC issues difficult.

As noted in #1650, it is unsettling when a program behaves
differently every time you run it.

While we're at it, remove unnecessary C_fix() call on the output of
time(NULL), as srand() is not a CHICKEN C function accepting fixnums,
but a native C function accepting regular integers.

Signed-off-by: Mario Domenech Goulart <mario@parenteses.org>