Don't merge syntactic environment for reexported macros (#1757)
Separate the syntactic exports into local and reexported macros,
so we can merge the syntactic environment into the local macros.

The original code would merge the reexporting module's syntactic
environment into every macro it exports, but this is incorrect for
reexported macros - those should stick with their original syntactic
environment, as it was in the defining module.

Signed-off-by: Evan Hanson <evhan@foldling.org>