Use merge-se, not append in ##sys#import to fix memory leak (#1772)
Using append on the current module (meta) environment in "import" will
extend the environment correctly, but it will grow the list.  Instead,
we want to replace add those bindings which are new (and possibly
replace existing bindings of the same name).  This is what merge-se
was made for.

This potentially slows down performance on import, but should be
relatively harmless.

Signed-off-by: felix <felix@call-with-current-continuation.org>