TARGET_SHARE_HOME should include "chicken".
TARGET_SHARE_HOME differs from INSTALL_SHARE_HOME by not having PROGRAM_PREFIX and PROGRAM_SUFFIX applied
(this is only necessary to distinguish multiple cross-chickens on the host), but
The chicken-specific destination directory should still be included on the target.
This patch also ensures "mac.r" is installed in the chicken-specific subdirectory of
PREFIX/share, which was previously incorrect, as reported by Lassi Kortela.

Signed-off-by: Peter Bex <peter@more-magic.net>