Ensure all ports are closed in tests when deleting the file
Under Windows, read-lines-test.scm and posix-tests.scm would
fail with "permission denied" upon deletion of the output file,
but that's because there was still an open file handle to it
(and on Windows you can't delete open files).

In read-lines-test.scm, it is my favorite Scheme footgun:
call-with-input-file doesn't close its port when the dynamic
extent is exited via other ways than a regular procedure return.

In posix-tests.scm it was a more mundane error - we opened a
file using the descriptor-returning POSIX procedure file-open,
but never closed the descriptor.

Signed-off-by: Mario Domenech Goulart <mario@parenteses.org>