Simplify library requirement processing
This is a "clarity" patch that simplifies the handling of
`##core#require` forms and tries to improve the naming of a couple of
related identifiers to keep things internally consistent.

First, rename the `core-unit?` procedure to `core-library?`, since it's
used to check all library IDs encountered during compilation and not
just units.

Then, use that procedure to check whether a library should be added to
the list of `required-libraries` directly within core.scm, rather than
returning multiple values from `##sys#process-require`. This separates
the logic for determining whether something is built-in from the
compiler's handling of `##core#require` forms a little more cleanly. If
something is a `core-library?` it should _not_ be added to the list,
else it should, the result being that `required-libraries` contains only
things that needs to be treated as an extension (and thus linked from
the eggs repository), and the interpreter doesn't have to care about any
of these decisions at all.

Then, because what we're building is really a list of _extensions_ that
need to be linked, rename `required-libraries` to `required-extensions`.

And finally, drop the list of `builtin-features`, which do not actually
need any special treatment from a library-loading point of view.

Signed-off-by: felix <felix@call-with-current-continuation.org>