Make "primitive modules" avoid triggering any code loading
This changes the way code loading is done, or rather _not_ done, for
so-called "primitive modules". Previously, these modules would be given
a library name that matches their module name, which would be passed to
the compiler, which would process a `##core#require` form only to decide
decide there's nothing to do (usually resulting in a `##core#undefined`
node). Really, this is useless work, since there is no library code to
load for any of these modules, and we know that up front. Instead, we
should just set the `library` field for these modules to false, to
indicate that they provide only syntax and don't need to load any code.
This is what's done by the change to `##sys#register-primitive-module`.

The two modules that are currently registered as "primitive modules" but
_do_ need to load code are chicken.foreign and chicken.condition, which
are handled as special cases in eval.scm. To keep the current behaviour,
we change these to use `##sys#register-core-module`, and explicitly set
their `library` field to the current value, i.e. the module name.

Note the changes in the patch do not only affect the modules defined
within modules.scm but also the chicken.csi module, for which the
`##sys#register-primitive-module` procedure is used in its import file.
As such, this fixes the problem of chicken.csi appearing in link files
because, as a primitive module, it will not create a `##core#require`
node and so will not end up in the list of `required-extensions`.

Also, drop the module alias for the srfi-88 library, which is already
defined (and more correctly) at modules.scm:1180.

Signed-off-by: felix <felix@call-with-current-continuation.org>