Update SRFI feature identifiers and documentation
Move all feature identifiers that are registered in chicken-syntax.scm
and expand.scm into the correct files and update the documentation that
relates to built-in features.

Add a section to "Included modules" that lists the SRFI modules that can
be imported. Also note that some SRFI identifiers _cannot_ be imported
because they are fully built-in and not defined by any module.

Signed-off-by: felix <felix@call-with-current-continuation.org>