Hopefully completely fix quoting hell in generated build  commands
Instead of using a mess of "qs*", "joins", "slashify" and "filelist"
and hoping for the best when generating chicken-do commands, use a
more principled method by way of a helper procedure to print the build
command.

This helper will receive *only* unquoted arguments: the list of
targets, list of dependencies and the build command line with flags as
a list.  It then calls "qs*" on all of these where needed and emits a
corresponding chicken-do line.

By doing it this way, it's much more obvious where quotation happens:
only in print-build-command, and never in the procedure that calls it.
For consistency, also change prepare-custom-command so that it accepts
an unquoted filename, so that quotation is delegated to it.

For now, leave the code that emits installation commands untouched.
We'll probably want to do the same for these though.

(by Peter Bex)

Signed-off-by: felix <felix@call-with-current-continuation.org>