Add a way to pass in already-quoted arguments to qs*
The librarian options are kept around in the binary as a foreign
string which may contain multiple flags separated by spaces.  These
should be taken verbatim by print-build-command.

The way we do this is by wrapping such verbatim snippets in a "raw"
record type and detecting these values in qs* and unpacking their
strings directly from the struct instead of quoting them.

Signed-off-by: felix <felix@call-with-current-continuation.org>