chicken-install: Consider two location layouts
When `location' is specified in setup.defaults, make chicken-install
consider two location layouts when looking for eggs:

* <location>/<egg>
* <location>/<egg>/<version>

This can be convenient for users who use clones of the
eggs-5-{all,latest} git repository of egg sources, or caches generated
by henrietta-cache.

As a bonus, support specification of versions when installing eggs
from a local directory.

Signed-off-by: felix <felix@call-with-current-continuation.org>